Time series analysis and sales data forecast
This project aims to perform a time series analysis and forecast the sales data using ARIMA and SARIMAX models to predict the sales. We first check if the present data is stationary or not using Dickey Fuller test then perform autoregression according to seasonality(12 months). Finally, we use ARIMA and SARIMAX to plot the forecast to crosscheck its accuracy with present data.